% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Optim.R
\name{print.Optim}
\alias{print.Optim}
\title{Print an Optim Object}
\usage{
\method{print}{Optim}(x, plain = FALSE, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{object of class "Optim"}

\item{plain}{select if you want enriched output mode (with colors and bold) or a plain output mode.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function prints an Optim object. It is a method for the generic function print of class "Optim".
}
