% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Square Error}
\usage{
RMSE(yhat, y, type.of = c("numeric", "text", "scalable"))
}
\arguments{
\item{yhat}{A predicted value vector}

\item{y}{A real value vector}

\item{type.of}{Type of response variable, either: \code{numeric} for the numerical response variables, \code{text} for the class response variables without growing relationship or \code{scalable} for the class response variables without growing relationship.}
}
\description{
RMSE is a commonly used error metric to measure the performance of regression models, but it is also possible to use it in a classification system.
The RMSE measures the standard deviation of the predictions from the ground-truth. This is the relationship between RMSE and classification.
}
